[Domain+Infra+App] Controllers+UseCase+Refactors

+++ src/main/java/com/fashionlike/proyecto_fashion_like/infra/persistence/repository/impl/PostRepositoryImpl.java
@@ -60,6 +60,9 @@ public class PostRepositoryImpl implements PostRepository {
     @Transactional
     @Override
     public Boolean deleteById(Integer id) {
+        if (!persistence.existsById(id)) {
+            return false;
+        }
         persistence.deleteById(id);
         return true;
     }
